tests/functional/requirements.txt
pytest==7.3.0
pytest-asyncio==0.21.0
allure-pytest==2.13.1
pydantic==1.10.7
python-dotenv==1.0.0
aiohttp==3.8.4
aiofiles==23.1.0